Fix : createHistory API Datatype

+++ fatman/app/src/main/java/com/project/fat/data/dto/CreateHistoryResponse.kt
@@ -1,8 +1,10 @@
 package com.project.fat.data.dto
 
+import java.time.LocalDateTime
+
 data class CreateHistoryResponse (
     val monsterNum: Long,
-    val distance: Long,
+    val distance: Double,
     val id: Long,
     val user: CreateHistoryUser,
     val date: String
